opt: move addPartialIndexPredicatesForTable to optbuilder/partial_index.go
Release note: None
