+++ pkg/sql/opt/optbuilder/select.go
@@ -688,80 +688,6 @@ func (b *Builder) addComputedColsForTable(tabMeta *opt.TableMeta) {
 	}
 }
 
-// addPartialIndexPredicatesForTable finds all partial indexes in the table and
-// adds their predicates to the table metadata (see
-// TableMeta.PartialIndexPredicates). The predicates are converted from strings
-// to ScalarExprs here.
-//
-// The predicates are used as "known truths" about table data. Any predicates
-// containing non-immutable operators are omitted.
-//
-// scan is an optional argument that is a Scan expression on the table. If scan
-// outputs all the ordinary columns in the table, we avoid constructing a new
-// scan. A scan and its logical properties are required in order to fully
-// normalize the partial index predicates.
-func (b *Builder) addPartialIndexPredicatesForTable(tabMeta *opt.TableMeta, scan memo.RelExpr) {
-	tab := tabMeta.Table
-
-	// Find the first partial index.
-	numIndexes := tab.IndexCount()
-	indexOrd := 0
-	for ; indexOrd < numIndexes; indexOrd++ {
-		if _, ok := tab.Index(indexOrd).Predicate(); ok {
-			break
-		}
-	}
-
-	// Return early if there are no partial indexes. Only partial indexes have
-	// predicates.
-	if indexOrd == numIndexes {
-		return
-	}
-
-	// Construct a scan as the tableScope expr so that logical properties of the
-	// scan can be used to fully normalize the index predicate.
-	tableScope := b.allocScope()
-	tableScope.appendOrdinaryColumnsFromTable(tabMeta, &tabMeta.Alias)
-
-	// If the optional scan argument was provided and it outputs all of the
-	// ordinary table columns, we use it as tableScope.expr. Otherwise, we must
-	// construct a new scan. Attaching a scan to tableScope.expr is required to
-	// fully normalize the partial index predicates with logical properties of
-	// the scan.
-	if scan != nil && tableScope.colSet().SubsetOf(scan.Relational().OutputCols) {
-		tableScope.expr = scan
-	} else {
-		tableScope.expr = b.factory.ConstructScan(&memo.ScanPrivate{
-			Table: tabMeta.MetaID,
-			Cols:  tableScope.colSet(),
-		})
-	}
-
-	// Skip to the first partial index we found above.
-	for ; indexOrd < numIndexes; indexOrd++ {
-		index := tab.Index(indexOrd)
-		pred, ok := index.Predicate()
-
-		// If the index is not a partial index, do nothing.
-		if !ok {
-			continue
-		}
-
-		expr, err := parser.ParseExpr(pred)
-		if err != nil {
-			panic(err)
-		}
-
-		// Build the partial index predicate as a memo.FiltersExpr and add it
-		// to the table metadata.
-		predExpr, err := b.buildPartialIndexPredicate(tableScope, expr, "index predicate")
-		if err != nil {
-			panic(err)
-		}
-		tabMeta.AddPartialIndexPredicate(indexOrd, &predExpr)
-	}
-}
-
 func (b *Builder) buildSequenceSelect(
 	seq cat.Sequence, seqName *tree.TableName, inScope *scope,
 ) (outScope *scope) {
